[JSONRPC] Add namespace query parameter support.
If the websocket query parameter 'ns' is found, its value will
become the callsign which will be automatically prepended to the
designator after parsing each JSONRPC message.  This will allow
clients to specify the callsign once when openning the websocket
and forgo including the callsign in the method field for every
json rpc request they make.

For firebolt project containing embedded callsigns, the designator
parsing is updated to respect the existance of multiple version numbers.